add support for grpc

+++ cleanup.py
@@ -56,48 +56,50 @@ def create_topic(admin_client, topic_name, num_partitions=1, replication_factor=
 if __name__ == '__main__':
     # Set up argument parser
     parser = argparse.ArgumentParser(description="Kafka topic management")
-    parser.add_argument('--topic', type=str, required=True, help='Kafka topic name')
+    parser.add_argument('--topic', type=str, required=False, help='Kafka topic name')
     parser.add_argument('--agents', type=int, required=False, help='Kafka topic name')
     parser.add_argument('--broker', type=str, required=False, help='Kafka Broker')
     parser.add_argument('--redis', type=str, required=False, help='Redis Service')
 
     # Parse command-line arguments
     args = parser.parse_args()
-    topic_name = args.topic
+    if args.topic:
+        topic_name = args.topic
 
-    bootstrap_servers = "localhost:19092"
-    if args.broker:
-        bootstrap_servers = args.broker
+        bootstrap_servers = "localhost:19092"
+        if args.broker:
+            bootstrap_servers = args.broker
 
-    admin_client = AdminClient({'bootstrap.servers': bootstrap_servers})
+        admin_client = AdminClient({'bootstrap.servers': bootstrap_servers})
 
-    if args.agents:
-        for x in range(1, args.agents + 1):
-            delete_topic(admin_client, f"{topic_name}-{x}")
-            delete_topic(admin_client, f"{topic_name}-hb-{x}")
-    else:
-        # Delete the topic
-        delete_topic(admin_client, topic_name)
-        delete_topic(admin_client, f"{topic_name}-hb")
+        if args.agents:
+            for x in range(1, args.agents + 1):
+                delete_topic(admin_client, f"{topic_name}-{x}")
+                delete_topic(admin_client, f"{topic_name}-hb-{x}")
+        else:
+            # Delete the topic
+            delete_topic(admin_client, topic_name)
+            delete_topic(admin_client, f"{topic_name}-hb")
 
-    time.sleep(1)
+        time.sleep(1)
 
-    if args.agents:
-        for x in range(1, args.agents + 1):
+        if args.agents:
+            for x in range(1, args.agents + 1):
+                # Create the topic
+                create_topic(admin_client, f"{topic_name}-{x}")
+                time.sleep(1)
+
+                create_topic(admin_client, f"{topic_name}-hb-{x}")
+                time.sleep(1)
+        else:
             # Create the topic
-            create_topic(admin_client, f"{topic_name}-{x}")
+            create_topic(admin_client, topic_name)
             time.sleep(1)
 
-            create_topic(admin_client, f"{topic_name}-hb-{x}")
+            create_topic(admin_client, f"{topic_name}-hb")
             time.sleep(1)
-    else:
-        # Create the topic
-        create_topic(admin_client, topic_name)
-        time.sleep(1)
 
-        create_topic(admin_client, f"{topic_name}-hb")
-        time.sleep(1)
-
-    redis_client = redis.StrictRedis(host=args.redis, port=6379, decode_responses=True)
-    task_repo = JobRepository(redis_client=redis_client)
-    task_repo.delete_all(key_prefix="*")
+    if args.redis:
+        redis_client = redis.StrictRedis(host=args.redis, port=6379, decode_responses=True)
+        task_repo = JobRepository(redis_client=redis_client)
+        task_repo.delete_all(key_prefix="*")
